finish did pallets
