+++ integration-tests/emulated/src/tests/peregrine/did_pallets/attestation.rs
@@ -4,24 +4,35 @@ use crate::{
 		para_chains::{AssetHubRococo, AssetHubRococoPallet, Peregrine},
 		relay_chains::Rococo,
 	},
-	tests::peregrine::did::utils::{
-		construct_xcm_message, get_asset_hub_sovereign_account, get_sibling_destination_peregrine,
+	tests::peregrine::did_pallets::utils::{
+		construct_xcm_message, create_mock_ctype, create_mock_did, get_asset_hub_sovereign_account,
+		get_sibling_destination_peregrine,
 	},
 };
-use did::did_details::DidVerificationKey;
-use frame_support::traits::fungible::hold::Inspect;
 use frame_support::{assert_ok, traits::fungible::Mutate};
 use parity_scale_codec::Encode;
-use rococo_runtime::System as RococoSystem;
 use runtime_common::{constants::KILT, AccountId, Balance};
-use xcm::{DoubleEncoded, VersionedXcm};
+use sp_core::H256;
+use xcm::VersionedXcm;
 use xcm_emulator::{assert_expected_events, OriginKind, Parachain, TestExt};
 
-fn get_xcm_message_create_did(origin_kind: OriginKind, withdraw_balance: Balance) -> VersionedXcm<()> {
+fn get_xcm_message_attestation_creation(
+	origin_kind: OriginKind,
+	withdraw_balance: Balance,
+	ctype_hash: H256,
+	claim_hash: H256,
+) -> VersionedXcm<()> {
 	let asset_hub_sovereign_account = get_asset_hub_sovereign_account();
 
-	let call: DoubleEncoded<()> = <Peregrine as Parachain>::RuntimeCall::Did(did::Call::create_from_account {
-		authentication_key: DidVerificationKey::Account(asset_hub_sovereign_account.clone()),
+	let call = <Peregrine as Parachain>::RuntimeCall::Did(did::Call::dispatch_as {
+		did_identifier: asset_hub_sovereign_account,
+		call: Box::new(<Peregrine as Parachain>::RuntimeCall::Attestation(
+			attestation::Call::add {
+				claim_hash,
+				ctype_hash,
+				authorization: None,
+			},
+		)),
 	})
 	.encode()
 	.into();
@@ -30,28 +41,39 @@ fn get_xcm_message_create_did(origin_kind: OriginKind, withdraw_balance: Balance
 }
 
 #[test]
-fn test_did_creation_from_asset_hub_successful() {
+fn test_attestation_creation_from_asset_hub_successful() {
 	MockNetworkRococo::reset();
 
 	let sudo_origin = <AssetHubRococo as Parachain>::RuntimeOrigin::root();
 
+	let ctype_hash_value = H256([0; 32]);
+	let claim_hash_value = H256([1; 32]);
+
 	let init_balance = KILT * 10;
 	let withdraw_balance = init_balance / 2;
 
-	let xcm = get_xcm_message_create_did(OriginKind::SovereignAccount, withdraw_balance);
+	let xcm_attestation_call = get_xcm_message_attestation_creation(
+		OriginKind::SovereignAccount,
+		withdraw_balance,
+		ctype_hash_value.clone(),
+		claim_hash_value.clone(),
+	);
+
 	let destination = get_sibling_destination_peregrine();
 
 	let asset_hub_sovereign_account = get_asset_hub_sovereign_account();
 
 	Peregrine::execute_with(|| {
+		create_mock_ctype(ctype_hash_value.clone());
+		create_mock_did();
 		<peregrine_runtime::Balances as Mutate<AccountId>>::set_balance(&asset_hub_sovereign_account, init_balance);
 	});
 
 	AssetHubRococo::execute_with(|| {
 		assert_ok!(<AssetHubRococo as AssetHubRococoPallet>::PolkadotXcm::send(
-			sudo_origin,
+			sudo_origin.clone(),
 			Box::new(destination.clone()),
-			Box::new(xcm.clone())
+			Box::new(xcm_attestation_call)
 		));
 
 		type RuntimeEvent = <AssetHubRococo as Parachain>::RuntimeEvent;
@@ -65,60 +87,60 @@ fn test_did_creation_from_asset_hub_successful() {
 
 	Peregrine::execute_with(|| {
 		type PeregrineRuntimeEvent = <Peregrine as Parachain>::RuntimeEvent;
+
 		assert_expected_events!(
 			Peregrine,
 			vec![
 				PeregrineRuntimeEvent::XcmpQueue(cumulus_pallet_xcmp_queue::Event::Success { .. }) => {},
-				PeregrineRuntimeEvent::Did(did::Event::DidCreated(account, did_identifier)) => {
-					account: account == &asset_hub_sovereign_account,
-					did_identifier:  did_identifier == &asset_hub_sovereign_account,
+				PeregrineRuntimeEvent::Attestation(attestation::Event::AttestationCreated { attester, claim_hash, authorization: _ , ctype_hash }) => {
+					attester: attester == &asset_hub_sovereign_account,
+					claim_hash: claim_hash == &claim_hash_value,
+					ctype_hash: ctype_hash == &ctype_hash_value,
 				},
 			]
-		);
-
-		let balance_on_hold = <<Peregrine as Parachain>::Balances as Inspect<AccountId>>::balance_on_hold(
-			&peregrine_runtime::RuntimeHoldReason::from(did::HoldReason::Deposit),
-			&asset_hub_sovereign_account,
-		);
-
-		assert_eq!(
-			balance_on_hold,
-			<peregrine_runtime::Runtime as did::Config>::BaseDeposit::get()
 		);
 	});
 
 	Rococo::execute_with(|| {
-		assert_eq!(RococoSystem::events().len(), 0);
+		assert_eq!(Rococo::events().len(), 0);
 	});
 }
 
 #[test]
-fn test_did_creation_from_asset_hub_unsuccessful() {
+fn test_attestation_creation_from_asset_hub_unsuccessful() {
+	let asset_hub_sovereign_account = get_asset_hub_sovereign_account();
 	let sudo_origin = <AssetHubRococo as Parachain>::RuntimeOrigin::root();
+	let destination = get_sibling_destination_peregrine();
+
+	let ctype_hash_value = H256([0; 32]);
+	let claim_hash_value = H256([1; 32]);
 
 	let init_balance = KILT * 100;
 	let withdraw_balance = init_balance / 2;
 
-	let asset_hub_sovereign_account = get_asset_hub_sovereign_account();
-	let destination = get_sibling_destination_peregrine();
+	let origin_kind_list = vec![OriginKind::Native, OriginKind::Superuser, OriginKind::Xcm];
 
-	let origin_kind_list = vec![OriginKind::Xcm, OriginKind::Superuser, OriginKind::Native];
-
-	for origin in origin_kind_list {
+	for origin_kind in origin_kind_list {
 		MockNetworkRococo::reset();
 
 		Peregrine::execute_with(|| {
+			create_mock_ctype(ctype_hash_value.clone());
+			create_mock_did();
 			<peregrine_runtime::Balances as Mutate<AccountId>>::set_balance(&asset_hub_sovereign_account, init_balance);
 		});
 
-		let xcm = get_xcm_message_create_did(origin, withdraw_balance);
+		let xcm_attestation_call = get_xcm_message_attestation_creation(
+			origin_kind,
+			withdraw_balance,
+			ctype_hash_value.clone(),
+			claim_hash_value.clone(),
+		);
 
-		//Send XCM message from AssetHub
 		AssetHubRococo::execute_with(|| {
 			assert_ok!(<AssetHubRococo as AssetHubRococoPallet>::PolkadotXcm::send(
 				sudo_origin.clone(),
 				Box::new(destination.clone()),
-				Box::new(xcm)
+				Box::new(xcm_attestation_call)
 			));
 
 			type RuntimeEvent = <AssetHubRococo as Parachain>::RuntimeEvent;
@@ -133,20 +155,17 @@ fn test_did_creation_from_asset_hub_unsuccessful() {
 		Peregrine::execute_with(|| {
 			type PeregrineRuntimeEvent = <Peregrine as Parachain>::RuntimeEvent;
 
-			let is_create_event_present = Peregrine::events().iter().any(|event| match event {
-				PeregrineRuntimeEvent::Did(did::Event::<peregrine_runtime::Runtime>::DidCreated(_, _)) => true,
+			let is_event_present = Peregrine::events().iter().any(|event| match event {
+				PeregrineRuntimeEvent::Did(did::Event::DidCallDispatched(_, _)) => true,
+				PeregrineRuntimeEvent::Attestation(attestation::Event::AttestationCreated { .. }) => true,
 				_ => false,
 			});
 
-			assert!(
-				!is_create_event_present,
-				"Create event for an unsupported origin is found"
-			);
+			assert!(!is_event_present);
 		});
 
-		// No event on the relaychain (message is meant for Peregrine)
 		Rococo::execute_with(|| {
-			assert_eq!(RococoSystem::events().len(), 0);
+			assert_eq!(Rococo::events().len(), 0);
 		});
 	}
 }
