+++ integration-tests/emulated/src/tests/peregrine/did_pallets/public_credentials.rs
@@ -4,34 +4,40 @@ use crate::{
 		para_chains::{AssetHubRococo, AssetHubRococoPallet, Peregrine},
 		relay_chains::Rococo,
 	},
-	tests::peregrine::did::utils::{
+	tests::peregrine::did_pallets::utils::{
 		construct_xcm_message, create_mock_ctype, create_mock_did, get_asset_hub_sovereign_account,
 		get_sibling_destination_peregrine,
 	},
 };
 use frame_support::{assert_ok, traits::fungible::Mutate};
+use kilt_asset_dids::AssetDid as AssetIdentifier;
 use parity_scale_codec::Encode;
-use rococo_runtime::System as RococoSystem;
 use runtime_common::{constants::KILT, AccountId, Balance};
 use sp_core::H256;
-use xcm::VersionedXcm;
+use sp_runtime::BoundedVec;
+use xcm::{DoubleEncoded, VersionedXcm};
 use xcm_emulator::{assert_expected_events, OriginKind, Parachain, TestExt};
 
-fn get_xcm_message_attestation_creation(
+fn get_xcm_message_add_public_credential(
 	origin_kind: OriginKind,
 	withdraw_balance: Balance,
 	ctype_hash: H256,
-	claim_hash: H256,
 ) -> VersionedXcm<()> {
 	let asset_hub_sovereign_account = get_asset_hub_sovereign_account();
 
-	let call = <Peregrine as Parachain>::RuntimeCall::Did(did::Call::dispatch_as {
+	let subject_id = AssetIdentifier::ether_currency();
+
+	let credential = public_credentials::mock::generate_base_public_credential_creation_op::<peregrine_runtime::Runtime>(
+		BoundedVec::try_from(subject_id.encode()).unwrap(),
+		ctype_hash,
+		Default::default(),
+	);
+
+	let call: DoubleEncoded<()> = <Peregrine as Parachain>::RuntimeCall::Did(did::Call::dispatch_as {
 		did_identifier: asset_hub_sovereign_account,
-		call: Box::new(<Peregrine as Parachain>::RuntimeCall::Attestation(
-			attestation::Call::add {
-				claim_hash,
-				ctype_hash,
-				authorization: None,
+		call: Box::new(<Peregrine as Parachain>::RuntimeCall::PublicCredentials(
+			public_credentials::Call::add {
+				credential: Box::new(credential),
 			},
 		)),
 	})
@@ -42,39 +48,31 @@ fn get_xcm_message_attestation_creation(
 }
 
 #[test]
-fn test_attestation_creation_from_asset_hub_successful() {
+fn test_create_public_credential_from_asset_hub() {
 	MockNetworkRococo::reset();
 
 	let sudo_origin = <AssetHubRococo as Parachain>::RuntimeOrigin::root();
-
+	let asset_hub_sovereign_account = get_asset_hub_sovereign_account();
 	let ctype_hash_value = H256([0; 32]);
-	let claim_hash_value = H256([1; 32]);
 
 	let init_balance = KILT * 10;
-	let withdraw_balance = init_balance / 2;
 
-	let xcm_attestation_call = get_xcm_message_attestation_creation(
-		OriginKind::SovereignAccount,
-		withdraw_balance,
-		ctype_hash_value.clone(),
-		claim_hash_value.clone(),
-	);
+	let xcm_claim_w3n_call =
+		get_xcm_message_add_public_credential(OriginKind::SovereignAccount, KILT, ctype_hash_value);
 
 	let destination = get_sibling_destination_peregrine();
 
-	let asset_hub_sovereign_account = get_asset_hub_sovereign_account();
-
 	Peregrine::execute_with(|| {
-		create_mock_ctype(ctype_hash_value.clone());
 		create_mock_did();
+		create_mock_ctype(ctype_hash_value.clone());
 		<peregrine_runtime::Balances as Mutate<AccountId>>::set_balance(&asset_hub_sovereign_account, init_balance);
 	});
 
 	AssetHubRococo::execute_with(|| {
 		assert_ok!(<AssetHubRococo as AssetHubRococoPallet>::PolkadotXcm::send(
-			sudo_origin.clone(),
-			Box::new(destination.clone()),
-			Box::new(xcm_attestation_call)
+			sudo_origin,
+			Box::new(destination),
+			Box::new(xcm_claim_w3n_call)
 		));
 
 		type RuntimeEvent = <AssetHubRococo as Parachain>::RuntimeEvent;
@@ -93,55 +91,49 @@ fn test_attestation_creation_from_asset_hub_successful() {
 			Peregrine,
 			vec![
 				PeregrineRuntimeEvent::XcmpQueue(cumulus_pallet_xcmp_queue::Event::Success { .. }) => {},
-				PeregrineRuntimeEvent::Attestation(attestation::Event::AttestationCreated { attester, claim_hash, authorization: _ , ctype_hash }) => {
-					attester: attester == &asset_hub_sovereign_account,
-					claim_hash: claim_hash == &claim_hash_value,
-					ctype_hash: ctype_hash == &ctype_hash_value,
+				PeregrineRuntimeEvent::Did(did::Event::DidCallDispatched(account, result)) => {
+					account: account == &asset_hub_sovereign_account,
+					result: result.is_ok(),
+				},
+				PeregrineRuntimeEvent::PublicCredentials(public_credentials::Event::CredentialStored{ subject_id: _, credential_id: _ }) => {
+
 				},
 			]
 		);
 	});
 
 	Rococo::execute_with(|| {
-		assert_eq!(RococoSystem::events().len(), 0);
+		assert_eq!(Rococo::events().len(), 0);
 	});
 }
 
 #[test]
-fn test_attestation_creation_from_asset_hub_unsuccessful() {
-	let asset_hub_sovereign_account = get_asset_hub_sovereign_account();
-	let sudo_origin = <AssetHubRococo as Parachain>::RuntimeOrigin::root();
-	let destination = get_sibling_destination_peregrine();
-
-	let ctype_hash_value = H256([0; 32]);
-	let claim_hash_value = H256([1; 32]);
-
-	let init_balance = KILT * 100;
-	let withdraw_balance = init_balance / 2;
-
+fn test_create_public_credential_from_asset_hub_unsuccessful() {
 	let origin_kind_list = vec![OriginKind::Native, OriginKind::Superuser, OriginKind::Xcm];
+
+	let sudo_origin = <AssetHubRococo as Parachain>::RuntimeOrigin::root();
+	let init_balance = KILT * 100;
+	let ctype_hash_value = H256([0; 32]);
+
+	let destination = get_sibling_destination_peregrine();
+	let asset_hub_sovereign_account = get_asset_hub_sovereign_account();
 
 	for origin_kind in origin_kind_list {
 		MockNetworkRococo::reset();
 
 		Peregrine::execute_with(|| {
-			create_mock_ctype(ctype_hash_value.clone());
 			create_mock_did();
+			create_mock_ctype(ctype_hash_value.clone());
 			<peregrine_runtime::Balances as Mutate<AccountId>>::set_balance(&asset_hub_sovereign_account, init_balance);
 		});
 
-		let xcm_attestation_call = get_xcm_message_attestation_creation(
-			origin_kind,
-			withdraw_balance,
-			ctype_hash_value.clone(),
-			claim_hash_value.clone(),
-		);
+		let xcm_claim_w3n_call = get_xcm_message_add_public_credential(origin_kind, KILT, ctype_hash_value);
 
 		AssetHubRococo::execute_with(|| {
 			assert_ok!(<AssetHubRococo as AssetHubRococoPallet>::PolkadotXcm::send(
 				sudo_origin.clone(),
 				Box::new(destination.clone()),
-				Box::new(xcm_attestation_call)
+				Box::new(xcm_claim_w3n_call.clone())
 			));
 
 			type RuntimeEvent = <AssetHubRococo as Parachain>::RuntimeEvent;
@@ -158,15 +150,15 @@ fn test_attestation_creation_from_asset_hub_unsuccessful() {
 
 			let is_event_present = Peregrine::events().iter().any(|event| match event {
 				PeregrineRuntimeEvent::Did(did::Event::DidCallDispatched(_, _)) => true,
-				PeregrineRuntimeEvent::Attestation(attestation::Event::AttestationCreated { .. }) => true,
+				PeregrineRuntimeEvent::DidLookup(pallet_did_lookup::Event::AssociationEstablished(_, _)) => true,
 				_ => false,
 			});
 
-			assert!(!is_event_present);
+			assert!(!is_event_present)
 		});
 
 		Rococo::execute_with(|| {
-			assert_eq!(RococoSystem::events().len(), 0);
+			assert_eq!(Rococo::events().len(), 0);
 		});
 	}
 }
