+++ integration-tests/emulated/src/tests/peregrine/did_pallets/unallowed_did_call.rs
@@ -4,13 +4,12 @@ use crate::{
 		para_chains::{AssetHubRococo, AssetHubRococoPallet, Peregrine},
 		relay_chains::Rococo,
 	},
-	tests::peregrine::did::utils::{
+	tests::peregrine::did_pallets::utils::{
 		construct_xcm_message, create_mock_did, get_asset_hub_sovereign_account, get_sibling_destination_peregrine,
 	},
 };
 use frame_support::{assert_ok, traits::fungible::Mutate};
 use parity_scale_codec::Encode;
-use rococo_runtime::System as RococoSystem;
 use runtime_common::{constants::KILT, AccountId, Balance};
 use xcm::{DoubleEncoded, VersionedXcm};
 use xcm_emulator::{assert_expected_events, OriginKind, Parachain, TestExt};
@@ -23,6 +22,24 @@ fn get_xcm_message_system_remark(origin_kind: OriginKind, withdraw_balance: Bala
 		call: Box::new(<Peregrine as Parachain>::RuntimeCall::System(
 			frame_system::Call::remark { remark: vec![] },
 		)),
+	})
+	.encode()
+	.into();
+
+	construct_xcm_message(origin_kind, withdraw_balance, call)
+}
+
+fn get_xcm_message_recursion(origin_kind: OriginKind, withdraw_balance: Balance) -> VersionedXcm<()> {
+	let asset_hub_sovereign_account = get_asset_hub_sovereign_account();
+
+	let call: DoubleEncoded<()> = <Peregrine as Parachain>::RuntimeCall::Did(did::Call::dispatch_as {
+		did_identifier: asset_hub_sovereign_account.clone(),
+		call: Box::new(<Peregrine as Parachain>::RuntimeCall::Did(did::Call::dispatch_as {
+			did_identifier: asset_hub_sovereign_account,
+			call: Box::new(<Peregrine as Parachain>::RuntimeCall::System(
+				frame_system::Call::remark { remark: vec![] },
+			)),
+		})),
 	})
 	.encode()
 	.into();
@@ -87,7 +104,69 @@ fn test_not_allowed_did_call() {
 		});
 
 		Rococo::execute_with(|| {
-			assert_eq!(RococoSystem::events().len(), 0);
+			assert_eq!(Rococo::events().len(), 0);
+		});
+	}
+}
+
+#[test]
+fn test_recursion_did_call() {
+	let origin_kind_list = vec![
+		OriginKind::Native,
+		OriginKind::Superuser,
+		OriginKind::Xcm,
+		OriginKind::SovereignAccount,
+	];
+
+	let sudo_origin = <AssetHubRococo as Parachain>::RuntimeOrigin::root();
+	let init_balance = KILT * 100;
+
+	let destination = get_sibling_destination_peregrine();
+	let asset_hub_sovereign_account = get_asset_hub_sovereign_account();
+
+	for origin_kind in origin_kind_list {
+		MockNetworkRococo::reset();
+
+		Peregrine::execute_with(|| {
+			create_mock_did();
+			<spiritnet_runtime::Balances as Mutate<AccountId>>::set_balance(&asset_hub_sovereign_account, init_balance);
+		});
+
+		let xcm_claim_w3n_call = get_xcm_message_recursion(origin_kind, KILT);
+
+		AssetHubRococo::execute_with(|| {
+			assert_ok!(<AssetHubRococo as AssetHubRococoPallet>::PolkadotXcm::send(
+				sudo_origin.clone(),
+				Box::new(destination.clone()),
+				Box::new(xcm_claim_w3n_call.clone())
+			));
+
+			type RuntimeEvent = <AssetHubRococo as Parachain>::RuntimeEvent;
+			assert_expected_events!(
+				AssetHubRococo,
+				vec![
+					RuntimeEvent::PolkadotXcm(pallet_xcm::Event::Sent { .. }) => {},
+				]
+			);
+		});
+
+		Peregrine::execute_with(|| {
+			type PeregrineRuntimeEvent = <Peregrine as Parachain>::RuntimeEvent;
+
+			// All calls should have [NoPermission] error
+			assert_expected_events!(
+				Peregrine,
+				vec![
+					PeregrineRuntimeEvent::XcmpQueue(cumulus_pallet_xcmp_queue::Event::Fail {
+						error: xcm::v3::Error::NoPermission,
+						..
+					}) => {},
+				]
+			);
+		});
+
+		Rococo::execute_with(|| {
+			assert_eq!(Rococo::events().len(), 0);
 		});
 	}
 }
