+++ integration-tests/emulated/src/tests/spiritnet/did_pallets/ctype.rs
@@ -1,0 +1,143 @@
+use crate::{
+	mock::{
+		network::MockNetworkPolkadot,
+		para_chains::{AssetHubPolkadot, AssetHubPolkadotPallet, Spiritnet},
+		relay_chains::Polkadot,
+	},
+	tests::spiritnet::did_pallets::utils::{
+		construct_xcm_message, create_mock_did, get_asset_hub_sovereign_account, get_sibling_destination_spiritnet,
+	},
+};
+use frame_support::{assert_ok, traits::fungible::Mutate};
+use parity_scale_codec::Encode;
+use runtime_common::{constants::KILT, AccountId, Balance};
+use xcm::{DoubleEncoded, VersionedXcm};
+use xcm_emulator::{assert_expected_events, OriginKind, Parachain, TestExt};
+
+fn get_xcm_message_ctype_creation(origin_kind: OriginKind, withdraw_balance: Balance) -> VersionedXcm<()> {
+	let asset_hub_sovereign_account = get_asset_hub_sovereign_account();
+
+	let call: DoubleEncoded<()> = <Spiritnet as Parachain>::RuntimeCall::Did(did::Call::dispatch_as {
+		did_identifier: asset_hub_sovereign_account,
+		call: Box::new(<Spiritnet as Parachain>::RuntimeCall::Ctype(ctype::Call::add {
+			ctype: b"{\"foo\": \"bar\"}".to_vec(),
+		})),
+	})
+	.encode()
+	.into();
+
+	construct_xcm_message(origin_kind, withdraw_balance, call)
+}
+
+#[test]
+fn test_ctype_creation_from_asset_hub_successful() {
+	MockNetworkPolkadot::reset();
+
+	let sudo_origin = <AssetHubPolkadot as Parachain>::RuntimeOrigin::root();
+
+	let init_balance = KILT * 10;
+
+	let xcm_ctype_call = get_xcm_message_ctype_creation(OriginKind::SovereignAccount, KILT);
+	let destination = get_sibling_destination_spiritnet();
+
+	let asset_hub_sovereign_account = get_asset_hub_sovereign_account();
+
+	Spiritnet::execute_with(|| {
+		create_mock_did();
+		<spiritnet_runtime::Balances as Mutate<AccountId>>::set_balance(&asset_hub_sovereign_account, init_balance);
+	});
+
+	AssetHubPolkadot::execute_with(|| {
+		assert_ok!(<AssetHubPolkadot as AssetHubPolkadotPallet>::PolkadotXcm::send(
+			sudo_origin,
+			Box::new(destination),
+			Box::new(xcm_ctype_call)
+		));
+
+		type RuntimeEvent = <AssetHubPolkadot as Parachain>::RuntimeEvent;
+		assert_expected_events!(
+			AssetHubPolkadot,
+			vec![
+				RuntimeEvent::PolkadotXcm(pallet_xcm::Event::Sent { .. }) => {},
+			]
+		);
+	});
+
+	Spiritnet::execute_with(|| {
+		type SpiritnetRuntimeEvent = <Spiritnet as Parachain>::RuntimeEvent;
+
+		assert_expected_events!(
+			Spiritnet,
+			vec![
+				SpiritnetRuntimeEvent::XcmpQueue(cumulus_pallet_xcmp_queue::Event::Success { .. }) => {},
+				SpiritnetRuntimeEvent::Did(did::Event::DidCallDispatched(account, result)) => {
+					account: account == &asset_hub_sovereign_account,
+					result: result.is_ok(),
+				},
+				SpiritnetRuntimeEvent::Ctype(ctype::Event::CTypeCreated(account, _)) => {
+					account: account == &asset_hub_sovereign_account,
+				},
+			]
+		);
+	});
+
+	Polkadot::execute_with(|| {
+		assert_eq!(Polkadot::events().len(), 0);
+	});
+}
+
+#[test]
+fn test_ctype_creation_from_asset_hub_unsuccessful() {
+	let sudo_origin = <AssetHubPolkadot as Parachain>::RuntimeOrigin::root();
+
+	let init_balance = KILT * 10;
+
+	let origin_kind_list = vec![OriginKind::Native, OriginKind::Superuser, OriginKind::Xcm];
+
+	let destination = get_sibling_destination_spiritnet();
+
+	let asset_hub_sovereign_account = get_asset_hub_sovereign_account();
+
+	for origin_kind in origin_kind_list {
+		MockNetworkPolkadot::reset();
+
+		Spiritnet::execute_with(|| {
+			create_mock_did();
+			<spiritnet_runtime::Balances as Mutate<AccountId>>::set_balance(&asset_hub_sovereign_account, init_balance);
+		});
+
+		let xcm_ctype_call = get_xcm_message_ctype_creation(origin_kind, KILT);
+
+		AssetHubPolkadot::execute_with(|| {
+			assert_ok!(<AssetHubPolkadot as AssetHubPolkadotPallet>::PolkadotXcm::send(
+				sudo_origin.clone(),
+				Box::new(destination.clone()),
+				Box::new(xcm_ctype_call)
+			));
+
+			type RuntimeEvent = <AssetHubPolkadot as Parachain>::RuntimeEvent;
+			assert_expected_events!(
+				AssetHubPolkadot,
+				vec![
+					RuntimeEvent::PolkadotXcm(pallet_xcm::Event::Sent { .. }) => {},
+				]
+			);
+		});
+
+		Spiritnet::execute_with(|| {
+			type SpiritnetRuntimeEvent = <Spiritnet as Parachain>::RuntimeEvent;
+
+			let is_event_present = Spiritnet::events().iter().any(|event| match event {
+				SpiritnetRuntimeEvent::Did(did::Event::DidCallDispatched(_, _)) => true,
+				SpiritnetRuntimeEvent::Ctype(ctype::Event::CTypeCreated(_, _)) => true,
+				_ => false,
+			});
+
+			assert!(!is_event_present);
+		});
+
+		Polkadot::execute_with(|| {
+			assert_eq!(Polkadot::events().len(), 0);
+		});
+	}
+}
