+++ integration-tests/emulated/src/tests/spiritnet/did_pallets/did.rs
@@ -1,25 +1,22 @@
-use crate::mock::{
-	network::MockNetworkPolkadot,
-	para_chains::{spiritnet, AssetHubPolkadot, AssetHubPolkadotPallet, Spiritnet},
-	relay_chains::Polkadot,
+use crate::{
+	mock::{
+		network::MockNetworkPolkadot,
+		para_chains::{AssetHubPolkadot, AssetHubPolkadotPallet, Spiritnet},
+		relay_chains::Polkadot,
+	},
+	tests::spiritnet::did_pallets::utils::{
+		construct_xcm_message, get_asset_hub_sovereign_account, get_sibling_destination_spiritnet,
+	},
 };
 use did::did_details::DidVerificationKey;
 use frame_support::traits::fungible::hold::Inspect;
 use frame_support::{assert_ok, traits::fungible::Mutate};
 use parity_scale_codec::Encode;
 use runtime_common::{constants::KILT, AccountId, Balance};
-use xcm::{v3::WeightLimit, DoubleEncoded, VersionedMultiLocation, VersionedXcm};
-use xcm_emulator::{
-	assert_expected_events, Here,
-	Instruction::{BuyExecution, Transact, WithdrawAsset},
-	Junction, Junctions, OriginKind, Parachain, ParentThen, TestExt, Weight, Xcm,
-};
+use xcm::{DoubleEncoded, VersionedXcm};
+use xcm_emulator::{assert_expected_events, OriginKind, Parachain, TestExt};
 
-fn get_asset_hub_sovereign_account() -> AccountId {
-	Spiritnet::sovereign_account_id_of(Spiritnet::sibling_location_of(AssetHubPolkadot::para_id()))
-}
-
-fn get_xcm_message(origin_kind: OriginKind, withdraw_balance: Balance) -> VersionedXcm<()> {
+fn get_xcm_message_create_did(origin_kind: OriginKind, withdraw_balance: Balance) -> VersionedXcm<()> {
 	let asset_hub_sovereign_account = get_asset_hub_sovereign_account();
 
 	let call: DoubleEncoded<()> = <Spiritnet as Parachain>::RuntimeCall::Did(did::Call::create_from_account {
@@ -28,24 +25,7 @@ fn get_xcm_message(origin_kind: OriginKind, withdraw_balance: Balance) -> Versio
 	.encode()
 	.into();
 
-	let require_weight_at_most = Weight::from_parts(10_000_600_000_000, 200_000_000_000);
-
-	VersionedXcm::from(Xcm(vec![
-		WithdrawAsset((Here, withdraw_balance).into()),
-		BuyExecution {
-			fees: (Here, withdraw_balance).into(),
-			weight_limit: WeightLimit::Unlimited,
-		},
-		Transact {
-			origin_kind,
-			require_weight_at_most,
-			call,
-		},
-	]))
-}
-
-fn get_destination() -> VersionedMultiLocation {
-	ParentThen(Junctions::X1(Junction::Parachain(spiritnet::PARA_ID))).into()
+	construct_xcm_message(origin_kind, withdraw_balance, call)
 }
 
 #[test]
@@ -57,21 +37,20 @@ fn test_did_creation_from_asset_hub_successful() {
 	let init_balance = KILT * 10;
 	let withdraw_balance = init_balance / 2;
 
-	let xcm = get_xcm_message(OriginKind::SovereignAccount, withdraw_balance);
-	let destination = get_destination();
+	let xcm = get_xcm_message_create_did(OriginKind::SovereignAccount, withdraw_balance);
+	let destination = get_sibling_destination_spiritnet();
 
 	let asset_hub_sovereign_account = get_asset_hub_sovereign_account();
-	// give the sovereign account of AssetHub some coins.
+
 	Spiritnet::execute_with(|| {
 		<spiritnet_runtime::Balances as Mutate<AccountId>>::set_balance(&asset_hub_sovereign_account, init_balance);
 	});
 
-	//Send XCM message from AssetHub
 	AssetHubPolkadot::execute_with(|| {
 		assert_ok!(<AssetHubPolkadot as AssetHubPolkadotPallet>::PolkadotXcm::send(
 			sudo_origin,
-			Box::new(destination),
-			Box::new(xcm)
+			Box::new(destination.clone()),
+			Box::new(xcm.clone())
 		));
 
 		type RuntimeEvent = <AssetHubPolkadot as Parachain>::RuntimeEvent;
@@ -88,28 +67,25 @@ fn test_did_creation_from_asset_hub_successful() {
 		assert_expected_events!(
 			Spiritnet,
 			vec![
+				SpiritnetRuntimeEvent::XcmpQueue(cumulus_pallet_xcmp_queue::Event::Success { .. }) => {},
 				SpiritnetRuntimeEvent::Did(did::Event::DidCreated(account, did_identifier)) => {
 					account: account == &asset_hub_sovereign_account,
 					did_identifier:  did_identifier == &asset_hub_sovereign_account,
 				},
-				SpiritnetRuntimeEvent::XcmpQueue(cumulus_pallet_xcmp_queue::Event::Success { .. }) => {},
 			]
 		);
 
-		// we also expect that the sovereignAccount of AssetHub has some coins now
 		let balance_on_hold = <<Spiritnet as Parachain>::Balances as Inspect<AccountId>>::balance_on_hold(
 			&spiritnet_runtime::RuntimeHoldReason::from(did::HoldReason::Deposit),
 			&asset_hub_sovereign_account,
 		);
 
-		// since a did is created, 2 of the free balance should now be on hold
 		assert_eq!(
 			balance_on_hold,
 			<spiritnet_runtime::Runtime as did::Config>::BaseDeposit::get()
 		);
 	});
 
-	// No event on the relaychain (message is meant for Spiritnet)
 	Polkadot::execute_with(|| {
 		assert_eq!(Polkadot::events().len(), 0);
 	});
@@ -117,24 +93,24 @@ fn test_did_creation_from_asset_hub_successful() {
 
 #[test]
 fn test_did_creation_from_asset_hub_unsuccessful() {
-	MockNetworkPolkadot::reset();
-
 	let sudo_origin = <AssetHubPolkadot as Parachain>::RuntimeOrigin::root();
 
-	let asset_hub_sovereign_account = get_asset_hub_sovereign_account();
-	let init_balance = KILT * 10;
+	let init_balance = KILT * 100;
 	let withdraw_balance = init_balance / 2;
 
-	let destination = get_destination();
+	let asset_hub_sovereign_account = get_asset_hub_sovereign_account();
+	let destination = get_sibling_destination_spiritnet();
+
 	let origin_kind_list = vec![OriginKind::Xcm, OriginKind::Superuser, OriginKind::Native];
 
 	for origin in origin_kind_list {
-		let xcm = get_xcm_message(origin, withdraw_balance);
+		MockNetworkPolkadot::reset();
 
-		// give the sovereign account of AssetHub some coins.
 		Spiritnet::execute_with(|| {
 			<spiritnet_runtime::Balances as Mutate<AccountId>>::set_balance(&asset_hub_sovereign_account, init_balance);
 		});
+
+		let xcm = get_xcm_message_create_did(origin, withdraw_balance);
 
 		//Send XCM message from AssetHub
 		AssetHubPolkadot::execute_with(|| {
@@ -156,15 +132,6 @@ fn test_did_creation_from_asset_hub_unsuccessful() {
 		Spiritnet::execute_with(|| {
 			type SpiritnetRuntimeEvent = <Spiritnet as Parachain>::RuntimeEvent;
 
-			// we still expect that the xcm message is send.
-			assert_expected_events!(
-				Spiritnet,
-				vec![
-					SpiritnetRuntimeEvent::XcmpQueue(cumulus_pallet_xcmp_queue::Event::Success { .. }) => {},
-				]
-			);
-
-			// ... but we also expect that the extrinsic will fail since it is no signed runtime origin.
 			let is_create_event_present = Spiritnet::events().iter().any(|event| match event {
 				SpiritnetRuntimeEvent::Did(did::Event::<spiritnet_runtime::Runtime>::DidCreated(_, _)) => true,
 				_ => false,
