+++ integration-tests/emulated/src/tests/spiritnet/did_pallets/mod.rs
@@ -1,0 +1,8 @@
+mod association;
+mod attestation;
+mod ctype;
+mod did;
+mod public_credentials;
+mod unallowed_did_call;
+mod utils;
+mod w3n;
