+++ integration-tests/emulated/src/tests/spiritnet/mod.rs
@@ -1,4 +1,4 @@
-mod did;
+mod did_pallets;
 mod reserve_transfers;
 mod sudo;
 mod teleportation;
